OSAC-1344: move validateOpenSSHPublicKey to shared ssh_validation.go
