+++ internal/servers/ssh_validation.go
@@ -1,0 +1,32 @@
+/*
+Copyright (c) 2025 Red Hat Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the
+License. You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
+language governing permissions and limitations under the License.
+*/
+
+package servers
+
+import (
+	"fmt"
+
+	"golang.org/x/crypto/ssh"
+)
+
+// validateOpenSSHPublicKey checks that the provided string is a valid OpenSSH public key.
+func validateOpenSSHPublicKey(key string) error {
+	_, _, _, rest, err := ssh.ParseAuthorizedKey([]byte(key))
+	if err != nil {
+		return fmt.Errorf("invalid OpenSSH public key: %w", err)
+	}
+	if len(rest) > 0 {
+		return fmt.Errorf("invalid OpenSSH public key: unexpected trailing content")
+	}
+	return nil
+}
